Create index.php
This is a very very stupid php script, but it is working. A very simple trial and error script to create a voucher and see the data in a HTML page, than you can remove the voucher file and when reloading the page it will auto create a new voucher.

Disclaimer: I am no coder. If someone want to alter, modify and clean up the code: PPPPPPPLEEEASSSSE !!!